refactor : directory 구조 변경
