+++ src/main/java/hotspot/user/outbox/notificationOutbox/infrastructure/NotificationOutboxEventJpaRepository.java
@@ -1,10 +1,10 @@
-package hotspot.user.outbox.infrastructure;
+package hotspot.user.outbox.notificationOutbox.infrastructure;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import hotspot.user.outbox.infrastructure.entity.NotificationOutboxEventEntity;
+import hotspot.user.outbox.notificationOutbox.infrastructure.entity.NotificationOutboxEventEntity;
 
 public interface NotificationOutboxEventJpaRepository extends JpaRepository<NotificationOutboxEventEntity, UUID> {
 }
