+++ src/test/java/hotspot/user/outbox/service/NotificationOutboxEventEventEntityAppenderTest.java
@@ -20,11 +20,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import hotspot.user.common.exception.ApplicationException;
 import hotspot.user.common.exception.code.OutboxErrorCode;
-import hotspot.user.outbox.infrastructure.NotificationOutboxEventJpaRepository;
-import hotspot.user.outbox.infrastructure.entity.NotificationOutboxEventEntity;
+import hotspot.user.outbox.notificationOutbox.infrastructure.NotificationOutboxEventJpaRepository;
+import hotspot.user.outbox.notificationOutbox.infrastructure.entity.NotificationOutboxEventEntity;
+import hotspot.user.outbox.notificationOutbox.service.NotificationOutboxEventAppender;
 
 @ExtendWith(MockitoExtension.class)
-class NotificationOutboxEventAppenderTest {
+class NotificationOutboxEventEventEntityAppenderTest {
 
     @Mock
     private NotificationOutboxEventJpaRepository outboxEventJpaRepository;
